bootcamp/requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2
djangorestframework==3.13.1
frozenlist==1.3.1
gunicorn==20.1.0
idna==3.4
multidict==6.0.2
Pillow==9.2.0
psycopg2-binary==2.9.3
pyaes==1.6.1
Pyrogram==2.0.59
PySocks==1.7.1
python-dotenv==0.20.0
pytz==2022.5
sqlparse==0.4.2
TgCrypto==1.2.4
yarl==1.8.1